Added dynamic memory allocation Added better build system for external binaries Fixed drive number mixups Made ramdisk driver use malloc

+++ src/drv/ramdisk.h
@@ -14,8 +14,8 @@
 
 diskStatus ramdisk_init(uint8_t drive);
 diskStatus ramdisk_status(uint8_t drive);
-diskResult ramdisk_write(uint8_t drive, const uint8_t *buff, uint32_t sector, uint32_t len);
-diskResult ramdisk_read(uint8_t drive, uint8_t *buff, uint32_t sector, uint32_t len);
+diskResult ramdisk_write(uint8_t drive, const uint8_t *buff, uint32_t sector, uint8_t len);
+diskResult ramdisk_read(uint8_t drive, uint8_t *buff, uint32_t sector, uint8_t len);
 diskResult ramdisk_ioctl(uint8_t drive, uint8_t cmd, void *buff);
 
 #endif
